Avoid deleting of journals during loading to database. JIRA: CXX-402
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@38437 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ scripts/common/check/check_make_win.sh
@@ -273,8 +273,10 @@ count_err=0
 count_absent=0
 configurations="$x_confs"
 
-rm -f "\$res_journal"
-rm -f "\$res_log"
+if ! is_db_load; then
+    rm -f "\$res_journal"
+    rm -f "\$res_log"
+fi
 
 ##  Run one test
